requirements.txt
python-dotenv
deepgram-sdk
livekit-plugins-openai
google-generativeai
tenacity
sqlalchemy